pyproject.toml
[tool.poetry]
name = "pochtman"
version = "0.0.2"
description = ""
authors = ["molodcovnik <molodcov-nikolayy@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.9.1"
aiosignal = "^1.3.1"
amqp = "^5.2.0"
asgiref = "^3.7.2"
attrs = "^23.2.0"
billiard = "^4.2.0"
celery = "^5.3.6"
certifi = "^2023.11.17"
cffi = "^1.16.0"
charset-normalizer = "^3.3.2"
click = "^8.1.7"
click-didyoumean = "^0.3.0"
click-plugins = "^1.1.1"
click-repl = "^0.3.0"
cryptography = "^42.0.2"
defusedxml = "^0.7.1"
django = "^5.0"
django-allauth = "^0.61.1"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.14.0"
frozenlist = "^1.4.1"
idna = "^3.6"
jinja2 = "^3.1.3"
kombu = "^5.3.5"
markupsafe = "^2.1.3"
multidict = "^6.0.4"
oauthlib = "^3.2.2"
prompt-toolkit = "^3.0.43"
psycopg2 = "^2.9.9"
pycparser = "^2.21"
pyjwt = "^2.8.0"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.1"
python3-openid = "^3.2.0"
pytz = "^2023.3.post1"
redis = "^5.0.1"
requests = "^2.31.0"
requests-oauthlib = "^1.3.1"
six = "^1.16.0"
sqlparse = "^0.4.4"
telepot = "^12.7"
tzdata = "^2024.1"
urllib3 = "^2.1.0"
vine = "^5.1.0"
wcwidth = "^0.2.13"
yarl = "^1.9.4"
drf-spectacular = "^0.27.2"
drf-spectacular-sidecar = "^2024.5.1"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
